+++ lib/rucaptcha/configuration.rb
@@ -1,5 +1,5 @@
 module RuCaptcha
   class Configuration
-    attr_accessor :width, :height, :font_size, :len
+    attr_accessor :width, :height, :font_size, :len, :implode
   end
 end