+++ lib/rucaptcha/controller_helpers.rb
@@ -7,12 +7,19 @@ module RuCaptcha
     end
 
     def generate_rucaptcha
-      session[:_rucaptcha] = SecureRandom.hex(RuCaptcha.config.len / 2)
-      return Captcha.create(session[:_rucaptcha])
+      session[:_rucaptcha] = random_rucaptcha_chars
+      return RuCaptcha::Captcha.create(session[:_rucaptcha])
+    end
+
+    def random_rucaptcha_chars
+      chars = SecureRandom.hex(RuCaptcha.config.len / 2).downcase
+      chars.gsub!(/[0ol1]/i, (rand(9) + 1).to_s)
+      chars
     end
 
     def verify_rucaptcha?(resource = nil)
-      right = params[:_rucaptcha].strip == session[:_rucaptcha]
+      params[:_rucaptcha] ||= ''
+      right = params[:_rucaptcha].downcase.strip == session[:_rucaptcha]
       if resource && resource.respond_to?(:errors)
         resource.errors.add(:base, t('rucaptcha.invalid')) unless right
       end