+++ CHANGELOG.md
@@ -1,4 +1,13 @@
-0.0.1
+0.1.2
+-----
+
+- No case sensitive;
+- Export config.implode;
+- Improve image color and style;
+- Don't generate chars in 'l,o,0,1'.
+- Render lower case chars on image.
+
+0.1.1
 -----
 
 - Include default validation I18n messages (en, zh-CN, zh-TW).